Update file paths of moved containers
